Split up the compiler pipeline for kernel and njit.
   - Introduces two compiler pipelines for numba_dpex, one
     for kernel and another for njit (auo-offload).
   - Kernel compilation from FunctionIR still uses the
     OffloadCompiler. (To be fixed later)
   - Kernels from Python code are now lowered using Numba's
     NativeLowering pass.
