+++ numba_dpex/core/pipelines/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: 2020 - 2022 Intel Corporation
+#
+# SPDX-License-Identifier: Apache-2.0
