+++ numba_dpex/tests/test_offload_diagnostics.py
@@ -36,27 +36,3 @@ class TestOffloadDiagnostics:
             dpex_config.OFFLOAD_DIAGNOSTICS = 0
             assert "Auto-offloading" in got.getvalue()
             assert "Device -" in got.getvalue()
-
-    def test_kernel(self):
-        @dpex.kernel
-        def parallel_sum(a, b, c):
-            i = dpex.get_global_id(0)
-            c[i] = a[i] + b[i]
-
-        global_size = 10
-        N = global_size
-
-        a = np.array(np.random.random(N), dtype=np.float32)
-        b = np.array(np.random.random(N), dtype=np.float32)
-        c = np.ones_like(a)
-
-        device = dpctl.SyclDevice("opencl:gpu")
-        with dpctl.device_context(device):
-            dpex_config.OFFLOAD_DIAGNOSTICS = 1
-
-            with captured_stdout() as got:
-                parallel_sum[global_size, dpex.DEFAULT_LOCAL_SIZE](a, b, c)
-
-            dpex_config.OFFLOAD_DIAGNOSTICS = 0
-            assert "Auto-offloading" in got.getvalue()
-            assert "Device -" in got.getvalue()
